git-svn-id: https://svn.code.sf.net/p/tikiwiki/code/trunk@26273 b456876b-0849-0410-b77d-98878d47e9d5

+++ lib/cache/cachelib.php
@@ -110,7 +110,7 @@ class Cachelib
 			}
 
 			while (false !== ($file = readdir($dir))) {
-				if (substr($file,0,1) == "." or $file == 'CVS' or $file == '.svn' or $file == "index.php" or $file == "README" or ($virtuals && in_array($file, $virtuals)) ) continue;
+				if (substr($file,0,1) == "." or $file == 'CVS' or $file == '.svn' or $file == "index.php" or $file == "README" or $file == "web.config" or ($virtuals && in_array($file, $virtuals)) ) continue;
 				if (is_dir($path."/".$file)) {
 					$this->erase_dir_content($path."/".$file);
 					rmdir($path."/".$file);
